+++ andes/interop/pandapower.py
@@ -471,7 +471,14 @@ def to_pandapower(ssa, ctrl=[], verify=True, tol=1e-6):
     Notes
     -----
     Handling of the following parameters:
-
+      - This interface tracks static power flow model in ANDES: `Bus`, `Line`, `PQ`, `Shunt`, `PV`, and `Slack`.
+        However, it does not track the dynamic models in ANDES, including but not limited to
+        `TurbineGov`, `SynGen`, and `Exciter`.
+      - The interface converts the ``Slack`` in ANDES to ``gen`` in pandapower rather than ``ext_gen``.
+      - MUST NOT verify power flow after initializing TDS in ANDES. ANDES does not allow running ``PFlow``
+        for systems with initialized TDS as it will break variable addressing.
+      - If you want to track dynamic model outputs in ANDES and feedback into pandapower,
+        you might need to manually transfer the results from ANDES to pandapower.
       - Generator cost is not included in the conversion. Use ``add_gencost()``
         to add cost data.
       - By default, all generators in ``ssp`` are controllable unless user-defined controllability
@@ -514,8 +521,26 @@ def to_pandapower(ssa, ctrl=[], verify=True, tol=1e-6):
 
 def _verify_pf(ssa, ssp, tol=1e-6):
     """
-    Verify power flow results.
+    Verify power flow results between ANDES and pandapower.
+
+    Parameters
+    ----------
+    tol : float
+        The tolerance of error.
+
+    Returns
+    -------
+    True
+        If the difference between the two results is within the tolerance
+    False
+        If the difference is larger than the tolerance, OR time-domain
+        simulation has been run in the ANDES system. ANDES is not able to run
+        power flow after initializing time-domain simulation.
     """
+
+    if ssa.TDS.initialized:
+        return False
+
     ssa.PFlow.run()
     pp.runpp(ssp)